Exception handling - delete

+++ src/main/java/com/cursospringboot/course/services/UserService.java
@@ -2,8 +2,11 @@ package com.cursospringboot.course.services;
 
 import com.cursospringboot.course.entities.User;
 import com.cursospringboot.course.repositories.UserRepository;
+import com.cursospringboot.course.services.exceptions.DatabaseException;
 import com.cursospringboot.course.services.exceptions.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -28,10 +31,17 @@ public class UserService {
         return repository.save(obj);
     }
 
-    public void delete(Long id){
-        repository.deleteById(id);
+    public void delete(Long id) {
+        try {
+            if (repository.existsById(id)) {
+                repository.deleteById(id);
+            } else {
+                throw new ResourceNotFoundException(id);
+            }
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException(e.getMessage());
+        }
     }
-
     public User update(Long id, User obj){
         User entity = repository.getReferenceById(id);
         updateData(entity, obj);
